fix: evaluate TEST_COVERAGE before to convert cov file

+++ .ci/scripts/unit-test.sh
@@ -5,10 +5,11 @@ source ./script/common.bash
 
 jenkins_setup
 
-export OUT_FILE="build/TEST-go-unit"
-
-go install -modfile=tools/go.mod github.com/t-yuki/gocover-cobertura
-
 make update
 mage goTestUnit
-gocover-cobertura < "${OUT_FILE}.cov" > "${OUT_FILE}.xml"
+
+OUT_FILE="build/TEST-go-unit"
+if [ -n "${TEST_COVERAGE}" ] && [ -f "${OUT_FILE}.cov" ]; then
+  go install -modfile=tools/go.mod github.com/t-yuki/gocover-cobertura
+  gocover-cobertura < "${OUT_FILE}.cov" > "${OUT_FILE}.xml"
+fi
